[command-compat]Update permission and signal handling
修改内容:
- 在 env 信号处理列表中直接使用 sigaction 的处理器字段进行状态比较。
- 将 sort 输出文件权限探测收敛为追加打开，避免额外写入标志干扰探测语义。

实现功能:
- 保持信号状态输出和输出文件可写性检测的兼容行为，减少平台类型和权限探测差异。

+++ crates/commands/env/src/env.rs
@@ -585,7 +585,7 @@ fn list_signal_handling() {
 
         let mut old_act: libc::sigaction = unsafe { std::mem::zeroed() };
         if unsafe { libc::sigaction(sig as libc::c_int, std::ptr::null(), &mut old_act) } == 0 {
-            let handler = old_act.sa_sigaction as usize;
+            let handler = old_act.sa_sigaction;
 
             let state_str = if handler == libc::SIG_IGN {
                 "IGNORE"
